agentcore-gateway/agent/requirements.txt
boto3>=1.35.0
bedrock-agentcore>=1.2.0
strands-agents>=0.1.0
